Refactor gui.py
Refactored into:
- main.py, which builds the GUI
- receive.py, which controls the receive tab
- send.py, which controls the send tab
- common.py, which contains common code

For #9
