+++ pyproject.toml
@@ -23,4 +23,4 @@ classifiers = [
 
 [project.scripts]
 pydiode = "pydiode.pydiode:main"
-pydiode-gui = "pydiode.gui:main"
+pydiode-gui = "pydiode.gui.main:main"
